move proxyablestore out of the calendar module
the proxyablestore did not belong in the calendar module.
it was moved to the core module.
